Refactor covenant naming to subscriptions
