+++ src/components/dashboard/SidebarContent.js
@@ -2,13 +2,11 @@ import React, {Component, PropTypes} from 'react';
 import SidebarSelector from './SidebarSelector';
 import BloodlinesSidebar from './BloodlinesSidebar';
 import RoasterSidebar from './RoasterSidebar';
-import CovenantSidebar from './CovenantSidebar';
 
 class SidebarContent extends Component {
     render() {
         const bloodlines = <BloodlinesSidebar location={this.props.location} />;
         const roaster = <RoasterSidebar location={this.props.location} />;
-        const covenant = <CovenantSidebar location={this.props.location}/>;
 
         const b = this.props.location;
         const d = '/dashboard/';
@@ -19,8 +17,8 @@ class SidebarContent extends Component {
                                      children={null} location={b} />
                     <SidebarSelector name="Bloodlines" to={d+"bloodlines"}
                         children={bloodlines} location={b} />
-                    <SidebarSelector name="Covenant" to={d+"covenant"}
-                        children={covenant} location={b} />
+                    <SidebarSelector name="Subscriptions" to={d+"subscriptions"}
+                        children={null} location={b} />
                     <SidebarSelector name="Roaster" to={d+"roaster"}
                                      children={roaster} location={b} />
                     <SidebarSelector name="Account Settings" to={d+"settings"}
