+++ src/components/dashboard/bloodlines/ReceiptContainer.js
@@ -28,6 +28,7 @@ class ReceiptContainer extends Component {
     }
 
     render() {
+        console.log(this.props.items);
         return (
             <div>
                 <ErrorMessage error={this.props.modify.error}/>
