+++ src/components/dashboard/subscriptions/SubscriptionContainer.js
@@ -1,7 +1,7 @@
-import React, { Component, PropTypes} from 'react';
-import { connect } from 'react-redux';
+import React, {Component, PropTypes} from 'react';
+import {connect} from 'react-redux';
 
-import { getAllSubscriptions, updateSubscription, deleteSubscription } from '../../../actions/covenantActions';
+import {getAllSubscriptions, updateSubscription, deleteSubscription } from '../../../actions/covenantActions';
 import ErrorMessage from '../../ErrorMessage';
 import SuccessMessage from '../../SuccessMessage';
 import SubscriptionList from './SubscriptionList';
@@ -25,7 +25,13 @@ class SubscriptionContainer extends Component {
 		if(reset) {
 			offset = 0;
 		}
-		dispatch(getAllSubscriptions(offset, 5)).then(this.nextPage.bind(this));
+		dispatch(getAllSubscriptions(offset, 10)).then(this.nextPage.bind(this));
+	}
+
+	nextPage() {
+		if(this.props.items.next && !this.props.items.fetching) {
+			this.update();
+		}
 	}
 
 	refresh() {
@@ -48,14 +54,8 @@ class SubscriptionContainer extends Component {
 		dispatch(deleteSubscription(id)).then(this.refresh.bind(this));
 	}
 
-	nextPage() {
-		if(this.props.items.next && !this.props.items.fetching) {
-			this.update();
-		}
-	}
-
 	render () {
-		console.log(this.props);
+		console.log(this.props.items);
 		return (
 			<div>
 				<h1> Subscriptions </h1>
